Separating test kit from integration tests namespace
