+++ core-integration-tests/src/test/java/com/jaspersoft/android/sdk/network/ReportOptionRestApiTest.java
@@ -24,11 +24,10 @@
 
 package com.jaspersoft.android.sdk.network;
 
-import com.jaspersoft.android.sdk.network.ReportOptionRestApi;
 import com.jaspersoft.android.sdk.network.entity.report.option.ReportOption;
-import com.jaspersoft.android.sdk.util.DummyTokenProvider;
-import com.jaspersoft.android.sdk.util.JrsMetadata;
-import com.jaspersoft.android.sdk.util.TestLogger;
+import com.jaspersoft.android.sdk.env.DummyTokenProvider;
+import com.jaspersoft.android.sdk.env.JrsMetadata;
+import com.jaspersoft.android.sdk.env.TestLogger;
 
 import org.junit.Before;
 import org.junit.Test;
