+++ core-integration-tests/src/test/java/com/jaspersoft/android/sdk/network/RepositoryRestApiTest.java
@@ -27,8 +27,8 @@ package com.jaspersoft.android.sdk.network;
 import com.jaspersoft.android.sdk.network.entity.resource.FolderLookup;
 import com.jaspersoft.android.sdk.network.entity.resource.ReportLookup;
 import com.jaspersoft.android.sdk.network.entity.resource.ResourceSearchResult;
-import com.jaspersoft.android.sdk.util.JrsEnvironmentRule;
-import com.jaspersoft.android.sdk.util.TestLogger;
+import com.jaspersoft.android.sdk.env.JrsEnvironmentRule;
+import com.jaspersoft.android.sdk.env.TestLogger;
 
 import org.junit.ClassRule;
 import org.junit.Test;
